Feature/add jwt auth (#8)
* config: install jwt lib.

* refactor: change api folder structure for better organization.

* feature: create login controller to generate JWT.

* feature: add route to get JWT and add secret to .env and .env.example. JWT generated successfully.

* feature: add middlewarewith guard function to protect routes with a JWT.

* add guard middleware to protect API routes.

* feature: add ownership to productController to apply JWT security.:

* docs: update swagger schema.

+++ api/controllers/apiProductController.js
@@ -1,9 +1,12 @@
 import { unlink } from 'node:fs/promises';
 import path from 'node:path';
-import Product from '../models/Product.js';
+import createError from 'http-errors';
+import Product from '../../models/Product.js';
 
 export const listProducts = async (req, res, next) => {
   try {
+    const userId = req.apiUserId;
+
     // http://localhost:3000/api/products?name=Nintendo
     const filterName = req.query.name;
     // http://localhost:3000/api/products?price=100-200
@@ -22,7 +25,7 @@ export const listProducts = async (req, res, next) => {
     const withCount = req.query.count === 'true';
 
     const filter = {
-      // owner: userId,
+      owner: userId,
     };
 
     if (filterName) {
@@ -58,7 +61,9 @@ export const listProducts = async (req, res, next) => {
 export const getOneProduct = async (req, res, next) => {
   try {
     const productId = req.params.productId;
-    const product = await Product.findById(productId);
+    const userId = req.apiUserId;
+
+    const product = await Product.findOne({ _id: productId, owner: userId });
 
     res.json({ results: product });
   } catch (error) {
@@ -69,9 +74,11 @@ export const getOneProduct = async (req, res, next) => {
 export const addNewProduct = async (req, res, next) => {
   try {
     const productData = req.body;
+    const userId = req.apiUserId;
 
     const product = new Product(productData);
     product.image = req.file?.filename;
+    product.owner = userId;
 
     const savedProduct = await product.save();
 
@@ -84,11 +91,15 @@ export const addNewProduct = async (req, res, next) => {
 export const updateProduct = async (req, res, next) => {
   try {
     const productId = req.params.productId;
+    const userId = req.apiUserId;
     const productData = req.body;
     productData.image = req.file?.filename;
 
-    const updatedProduct = await Product.findByIdAndUpdate(
-      productId,
+    const updatedProduct = await Product.findOneAndUpdate(
+      {
+        _id: productId,
+        owner: userId,
+      },
       productData,
       { new: true },
     );
@@ -102,7 +113,24 @@ export const updateProduct = async (req, res, next) => {
 export const deleteProduct = async (req, res, next) => {
   try {
     const productId = req.params.productId;
+    const userId = req.apiUserId;
+
     const product = await Product.findById(productId);
+
+    if (!product) {
+      console.warning(
+        `WARNING: user ${userId} is trying to delete non existing products.`,
+      );
+      return next(createError(404));
+    }
+
+    if (product.owner.toString() !== userId) {
+      console.warning(
+        `WARNING: user ${userId} is trying to delete products from other users.`,
+      );
+      return next(createError(401));
+    }
+
     if (product.image) {
       await unlink(
         path.join(
